Refactor : controller
