+++ delivery-relay-service/src/main/java/com/inbobwetrust/controller/HealthCheckController.java
@@ -1,4 +1,4 @@
-package com.inbobwetrust.relay;
+package com.inbobwetrust.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
