+++ delivery-relay-service/src/test/java/com/inbobwetrust/relay/RelayControllerTest.java
@@ -3,12 +3,15 @@ package com.inbobwetrust.relay;
 import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.when;
 
-import com.inbobwetrust.relay.domain.Delivery;
-import com.inbobwetrust.relay.domain.DeliveryStatus;
-import com.inbobwetrust.relay.domain.ReceiverType;
-import com.inbobwetrust.relay.domain.RelayRequest;
+import com.inbobwetrust.controller.RelayController;
+import com.inbobwetrust.domain.Delivery;
+import com.inbobwetrust.domain.DeliveryStatus;
+import com.inbobwetrust.domain.ReceiverType;
+import com.inbobwetrust.domain.RelayRequest;
 import java.time.LocalDateTime;
 import java.util.stream.Stream;
+
+import com.inbobwetrust.repository.RelayRepository;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -30,7 +33,8 @@ class RelayControllerTest {
 
   @Autowired WebTestClient testClient;
 
-  @MockBean RelayRepository relayRepository;
+  @MockBean
+  RelayRepository relayRepository;
 
   private final String BASE_URL = "/relay/v1";
